requirements.txt
Flask==0.12.2
boto3==1.9.78
numpy==1.14.0
Keras==2.1.6
requests==2.18.4
protobuf==3.6.1
google_api_python_client==1.7.7
